Fixed NMCheckStringName bug
This function was added to check wave names, but did not check if wave
was full path. Fixed this bug. The path and wave name are first
seperated and then the wave name is checked.

Added new compile #if statement to detect if NIDAQ XOP exists. This
allows NM to compile if NIDAQ XOP does not exist.