Fixed Pulse Gen SinZap bug
The linear sine zap/sweep function was missing a factor of 2. This bug created a sweep function that ended with a frequency half what it should have been. Also, the window over which the zap/sweep was computed was incorrect. This was changed to aow.width. Added new "type" parameter which allows one to choose between a linear sweep (type=0) and exponential sweep (type=1).